Remove ReaderDetailHeaderView
